Moved all the SB_Servo library dependencies into the dependencies folder. Makes it easier for someone to easily pull down the repo and start working
